+++ CHANGES.rst
@@ -27,12 +27,16 @@ Release history
 - It is now possible to slice the ``pre`` neurons in a neuron->neuron
   connection.
   (`#226 <https://github.com/nengo/nengo-loihi/pull/226>`__)
+- Connections now support ``Sparse`` transforms.
+  (`#240 <https://github.com/nengo/nengo-loihi/pull/240>`__)
 
 **Changed**
 
 - Connections from neurons with scalar transforms are now sparse internally.
   This allows much larger neuron->neuron connections with scalar transforms.
   (`#226 <https://github.com/nengo/nengo-loihi/pull/226>`__)
+- The ``scipy`` package is now required to run Nengo Loihi.
+  (`#240 <https://github.com/nengo/nengo-loihi/pull/240>`__)
 
 0.8.0 (June 23, 2019)
 =====================